Remove PacketFunction, now merged into the packet states.
