+++ src/main/java/org/spongepowered/common/event/tracking/phase/packet/AttackEntityPacketState.java
@@ -24,14 +24,37 @@
  */
 package org.spongepowered.common.event.tracking.phase.packet;
 
+import net.minecraft.entity.item.EntityItem;
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.network.Packet;
 import net.minecraft.network.play.client.CPacketUseEntity;
+import net.minecraft.world.WorldServer;
+import org.apache.logging.log4j.Level;
+import org.spongepowered.api.entity.Entity;
+import org.spongepowered.api.event.SpongeEventFactory;
+import org.spongepowered.api.event.cause.Cause;
 import org.spongepowered.api.event.cause.NamedCause;
+import org.spongepowered.api.event.cause.entity.spawn.EntitySpawnCause;
+import org.spongepowered.api.event.item.inventory.DropItemEvent;
 import org.spongepowered.api.item.inventory.ItemStack;
+import org.spongepowered.api.world.World;
+import org.spongepowered.asm.util.PrettyPrinter;
+import org.spongepowered.common.SpongeImpl;
+import org.spongepowered.common.entity.EntityUtil;
 import org.spongepowered.common.event.InternalNamedCauses;
+import org.spongepowered.common.event.tracking.ItemDropData;
 import org.spongepowered.common.event.tracking.PhaseContext;
+import org.spongepowered.common.event.tracking.TrackingUtil;
 import org.spongepowered.common.item.inventory.util.ItemStackUtil;
+import org.spongepowered.common.registry.type.event.InternalSpawnTypes;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
+import java.util.stream.Collectors;
 
 import javax.annotation.Nullable;
 
@@ -64,5 +87,99 @@ final class AttackEntityPacketState extends BasicPacketState {
     @Override
     public boolean tracksEntitySpecificDrops() {
         return true;
+    }
+
+    @Override
+    public void unwind(Packet<?> packet, EntityPlayerMP player, PhaseContext context) {
+        final CPacketUseEntity useEntityPacket = (CPacketUseEntity) packet;
+        final net.minecraft.entity.Entity entity = useEntityPacket.getEntityFromWorld(player.world);
+        if (entity == null) {
+            // Something happened?
+            return;
+        }
+        //final Optional<ItemStack> itemStack = context.firstNamed(InternalNamedCauses.Packet.ITEM_USED, ItemStack.class);
+        final World spongeWorld = EntityUtil.getSpongeWorld(player);
+        EntityUtil.toMixin(entity).setNotifier(player.getUniqueID());
+
+        context.getCapturedItemsSupplier()
+                .ifPresentAndNotEmpty(items -> {
+                    // For destruction, this should be empty, however, some times, it may not be?
+                    final PrettyPrinter printer = new PrettyPrinter(60);
+                    printer.add("Processing Attack Entity").centre().hr();
+                    printer.add("There are some captured items after the entity was destructed!");
+                    printer.addWrapped(60, "%s : %s", "Items captured", items);
+                    printer.add("Stacktrace:");
+                    printer.add(new Exception("Stack trace"));
+                    printer.trace(System.err, SpongeImpl.getLogger(), Level.TRACE);
+                });
+        context.getCapturedBlockSupplier()
+                .ifPresentAndNotEmpty(blocks ->
+                        TrackingUtil.processBlockCaptures(blocks, this, context));
+        context.getCapturedEntityDropSupplier().ifPresentAndNotEmpty(map -> {
+            for (Map.Entry<UUID, Collection<ItemDropData>> entry : map.asMap().entrySet()) {
+                final UUID key = entry.getKey();
+                final Optional<Entity> affectedEntity = spongeWorld.getEntity(key);
+                if (!affectedEntity.isPresent()) {
+                    continue;
+                }
+                final Collection<ItemDropData> itemStacks = entry.getValue();
+                if (itemStacks.isEmpty()) {
+                    return;
+                }
+                final List<ItemDropData> items = new ArrayList<>();
+                items.addAll(itemStacks);
+
+                if (!items.isEmpty()) {
+                    final List<Entity> itemEntities = items.stream()
+                            .map(data -> data.create(((WorldServer) player.world)))
+                            .map(EntityUtil::fromNative)
+                            .collect(Collectors.toList());
+                    final Cause cause = Cause.source(EntitySpawnCause.builder()
+                            .entity(affectedEntity.get())
+                            .type(InternalSpawnTypes.DROPPED_ITEM)
+                            .build()
+                    )
+                            .named(NamedCause.of("Attacker", player))
+                            .build();
+                    final DropItemEvent.Destruct
+                            destruct =
+                            SpongeEventFactory.createDropItemEventDestruct(cause, itemEntities);
+                    SpongeImpl.postEvent(destruct);
+                    if (!destruct.isCancelled()) {
+                        PacketPhaseUtil.processSpawnedEntities(player, destruct);
+                    }
+                }
+            }
+        });
+        context.getCapturedEntityItemDropSupplier().ifPresentAndNotEmpty(map -> {
+            for (Map.Entry<UUID, Collection<EntityItem>> entry : map.asMap().entrySet()) {
+                final UUID key = entry.getKey();
+                final Optional<Entity> attackedEntities = spongeWorld.getEntity(key);
+                if (!attackedEntities.isPresent()) {
+                    continue;
+                }
+                final List<Entity> items = entry.getValue().stream().map(EntityUtil::fromNative).collect(Collectors.toList());
+                final Cause cause = Cause.source(EntitySpawnCause.builder()
+                        .entity(EntityUtil.fromNative(player))
+                        .type(InternalSpawnTypes.DROPPED_ITEM)
+                        .build()
+                )
+                        .named(NamedCause.of("Attacker", player))
+                        .build();
+                final DropItemEvent.Destruct
+                        destruct =
+                        SpongeEventFactory.createDropItemEventDestruct(cause, items);
+                SpongeImpl.postEvent(destruct);
+                if (!destruct.isCancelled()) {
+                    PacketPhaseUtil.processSpawnedEntities(player, destruct);
+                }
+            }
+        });
+
+
+        context.getCapturedBlockSupplier()
+                .ifPresentAndNotEmpty(snapshots ->
+                        TrackingUtil.processBlockCaptures(snapshots, this, context)
+                );
     }
 }
