+++ src/main/java/org/spongepowered/common/event/tracking/phase/packet/BasicInventoryPacketState.java
@@ -26,75 +26,14 @@ package org.spongepowered.common.event.tracking.phase.packet;
 
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.network.Packet;
-import org.spongepowered.api.data.Transaction;
-import org.spongepowered.api.entity.Entity;
-import org.spongepowered.api.event.cause.Cause;
-import org.spongepowered.api.event.item.inventory.ClickInventoryEvent;
-import org.spongepowered.api.item.inventory.Container;
-import org.spongepowered.api.item.inventory.ItemStackSnapshot;
-import org.spongepowered.api.item.inventory.transaction.SlotTransaction;
 import org.spongepowered.common.event.tracking.PhaseContext;
 import org.spongepowered.common.interfaces.IMixinContainer;
 
-import java.util.List;
-
-import javax.annotation.Nullable;
-
-public class BasicInventoryPacketState extends BasicPacketState {
-
-
-    /**
-     * Flags we care about
-     */
-    final int stateId;
-
-    /**
-     * Mask for flags we care about, the caremask if you will
-     */
-    final int stateMask;
-
-    /**
-     * Don't care about anything
-     */
-    public BasicInventoryPacketState() {
-        this(0, PacketPhase.MASK_NONE);
-    }
-
-    /**
-     * We care a lot
-     *
-     * @param stateId state
-     */
-    public BasicInventoryPacketState(int stateId) {
-        this(stateId, PacketPhase.MASK_ALL);
-    }
-
-    /**
-     * We care about some things
-     *
-     * @param stateId flags we care about
-     * @param stateMask caring mask
-     */
-    public BasicInventoryPacketState(int stateId, int stateMask) {
-        this.stateId = stateId & stateMask;
-        this.stateMask = stateMask;
-    }
+public abstract class BasicInventoryPacketState extends BasicPacketState {
 
     @Override
     public boolean doBlockCapturing() {
         return false;
-    }
-
-
-    @Nullable
-    public ClickInventoryEvent createInventoryEvent(EntityPlayerMP playerMP, Container openContainer, Transaction<ItemStackSnapshot> transaction,
-            List<SlotTransaction> slotTransactions, List<Entity> capturedEntities, Cause cause, int usedButton) {
-        return null;
-    }
-
-    @Override
-    public boolean matches(int packetState) {
-        return this.stateMask != PacketPhase.MASK_NONE && ((packetState & this.stateMask & this.stateId) == (packetState & this.stateMask));
     }
 
     @Override
