+++ src/main/java/org/spongepowered/common/event/tracking/phase/packet/IgnoredPacketState.java
@@ -24,7 +24,15 @@
  */
 package org.spongepowered.common.event.tracking.phase.packet;
 
+import net.minecraft.entity.player.EntityPlayerMP;
+import net.minecraft.network.Packet;
+import org.spongepowered.common.event.tracking.PhaseContext;
+
 final class IgnoredPacketState extends BasicPacketState {
+
+    @Override
+    public void unwind(Packet<?> packet, EntityPlayerMP player, PhaseContext context) {
+    }
 
     @Override
     public boolean ignoresItemPreMerges() {
