+++ src/main/java/org/spongepowered/common/event/tracking/phase/packet/InteractEntityPacketState.java
@@ -27,11 +27,31 @@ package org.spongepowered.common.event.tracking.phase.packet;
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.network.Packet;
 import net.minecraft.network.play.client.CPacketUseEntity;
+import org.spongepowered.api.block.BlockSnapshot;
+import org.spongepowered.api.entity.Entity;
+import org.spongepowered.api.entity.living.player.Player;
+import org.spongepowered.api.event.SpongeEventFactory;
+import org.spongepowered.api.event.cause.Cause;
 import org.spongepowered.api.event.cause.NamedCause;
+import org.spongepowered.api.event.cause.entity.spawn.EntitySpawnCause;
+import org.spongepowered.api.event.entity.SpawnEntityEvent;
 import org.spongepowered.api.item.inventory.ItemStack;
+import org.spongepowered.api.world.World;
+import org.spongepowered.asm.util.PrettyPrinter;
+import org.spongepowered.common.SpongeImpl;
+import org.spongepowered.common.entity.EntityUtil;
 import org.spongepowered.common.event.InternalNamedCauses;
+import org.spongepowered.common.event.tracking.ItemDropData;
 import org.spongepowered.common.event.tracking.PhaseContext;
+import org.spongepowered.common.event.tracking.TrackingUtil;
 import org.spongepowered.common.item.inventory.util.ItemStackUtil;
+import org.spongepowered.common.registry.type.event.InternalSpawnTypes;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
+import java.util.stream.Collectors;
 
 import javax.annotation.Nullable;
 
@@ -76,5 +96,72 @@ final class InteractEntityPacketState extends BasicPacketState {
     @Override
     public boolean doesCaptureEntityDrops() {
         return true;
+    }
+
+    @Override
+    public void unwind(Packet<?> packet, EntityPlayerMP player, PhaseContext context) {
+        final CPacketUseEntity useEntityPacket = (CPacketUseEntity) packet;
+        final net.minecraft.entity.Entity entity = useEntityPacket.getEntityFromWorld(player.world);
+        if (entity == null) {
+            // Something happened?
+            return;
+        }
+        //final Optional<ItemStack> itemStack = context.firstNamed(InternalNamedCauses.Packet.ITEM_USED, ItemStack.class);
+        final World spongeWorld = EntityUtil.getSpongeWorld(player);
+        EntityUtil.toMixin(entity).setNotifier(player.getUniqueID());
+
+        context.getCapturedBlockSupplier()
+                .ifPresentAndNotEmpty(blocks -> {
+                    final PrettyPrinter printer = new PrettyPrinter(80);
+                    printer.add("Processing Interact Entity").centre().hr();
+                    printer.add("The blocks captured are:");
+                    for (BlockSnapshot blockSnapshot : blocks) {
+                        printer.add("  Block: %s", blockSnapshot);
+                    }
+                    printer.trace(System.err);
+                });
+        context.getCapturedEntitySupplier()
+                .ifPresentAndNotEmpty(entities -> {
+                    final PrettyPrinter printer = new PrettyPrinter(80);
+                    printer.add("Processing Interact Entity").centre().hr();
+                    printer.add("The entities captured are:");
+                    for (Entity capturedEntity : entities) {
+                        printer.add("  Entity: %s", capturedEntity);
+                    }
+                    printer.trace(System.err);
+                });
+        context.getCapturedItemsSupplier().ifPresentAndNotEmpty(entities -> {
+            final List<Entity> items = entities.stream().map(EntityUtil::fromNative).collect(Collectors.toList());
+            final Cause cause = Cause.source(EntitySpawnCause.builder()
+                    .entity((Player) player)
+                    .type(InternalSpawnTypes.PLACEMENT)
+                    .build()
+            ).build();
+            SpawnEntityEvent event = SpongeEventFactory.createSpawnEntityEvent(cause, items);
+            SpongeImpl.postEvent(event);
+            if (!event.isCancelled()) {
+                PacketPhaseUtil.processSpawnedEntities(player, event);
+
+            }
+        });
+        context.getCapturedEntityDropSupplier()
+                .ifPresentAndNotEmpty(map -> {
+                    final PrettyPrinter printer = new PrettyPrinter(80);
+                    printer.add("Processing Interact Entity").centre().hr();
+                    printer.add("The item stacks captured are: ");
+
+                    for (Map.Entry<UUID, Collection<ItemDropData>> entry : map.asMap().entrySet()) {
+                        printer.add("  - Entity with UUID: %s", entry.getKey());
+                        for (ItemDropData stack : entry.getValue()) {
+                            printer.add("    - %s", stack);
+                        }
+                    }
+                    printer.trace(System.err);
+                });
+
+        context.getCapturedBlockSupplier()
+                .ifPresentAndNotEmpty(snapshots ->
+                        TrackingUtil.processBlockCaptures(snapshots, this, context)
+                );
     }
 }
