+++ src/main/java/org/spongepowered/common/event/tracking/phase/packet/drag/DragInventoryStartState.java
@@ -24,7 +24,19 @@
  */
 package org.spongepowered.common.event.tracking.phase.packet.drag;
 
+import net.minecraft.entity.player.EntityPlayerMP;
+import org.spongepowered.api.data.Transaction;
+import org.spongepowered.api.entity.Entity;
+import org.spongepowered.api.event.cause.Cause;
+import org.spongepowered.api.event.item.inventory.ClickInventoryEvent;
+import org.spongepowered.api.item.inventory.Container;
+import org.spongepowered.api.item.inventory.ItemStackSnapshot;
+import org.spongepowered.api.item.inventory.transaction.SlotTransaction;
 import org.spongepowered.common.event.tracking.phase.packet.PacketPhase;
+
+import java.util.List;
+
+import javax.annotation.Nullable;
 
 public final class DragInventoryStartState extends NamedInventoryState {
 
@@ -32,4 +44,10 @@ public final class DragInventoryStartState extends NamedInventoryState {
         super(name, PacketPhase.MODE_DRAG | buttonId | PacketPhase.DRAG_STATUS_STARTED | PacketPhase.CLICK_OUTSIDE_WINDOW, PacketPhase.MASK_DRAG);
     }
 
+    @Nullable
+    @Override
+    public ClickInventoryEvent createInventoryEvent(EntityPlayerMP playerMP, Container openContainer, Transaction<ItemStackSnapshot> transaction,
+            List<SlotTransaction> slotTransactions, List<Entity> capturedEntities, Cause cause, int usedButton) {
+        return null;
+    }
 }
